Added Orders

+++ HVAC-Shop.Core/DTO/OrderDto.cs
@@ -8,13 +8,13 @@ namespace HVAC_Shop.Core.DTO
         public required string BuyerEmail { get; set; }
         public required ShippingAddress Address { get; set; }
         public DateTime OrderDate { get; set; }
-        public long Subtotal { get; set; }
-        public long DeliveryFee { get; set; }
-        public long Discount { get; set; }
+        public decimal Subtotal { get; set; }
+        public decimal DeliveryFee { get; set; }
+        public decimal Discount { get; set; }
         public OrderStatus Status { get; set; }
         public List<OrderItemDto> OrderItems { get; set; } = [];
         public required string PaymentIntentId { get; set; }
         public required PaymentSummery PaymentSummary { get; set; }
-        public long Total { get; set; }
+        public decimal Total { get; set; }
     }
 }
